正式版本

+++ src/main/java/me/fullidle/customrankingpapi/customrankingpapi/PlayerValue.java
@@ -2,20 +2,19 @@ package me.fullidle.customrankingpapi.customrankingpapi;
 
 import lombok.Getter;
 import org.bukkit.OfflinePlayer;
-import org.jetbrains.annotations.NotNull;
 
 @Getter
 public class PlayerValue implements Comparable<PlayerValue>{
     private final OfflinePlayer player;
-    private final Integer value;
+    private final Double value;
 
-    public PlayerValue(OfflinePlayer player,Integer value){
+    public PlayerValue(OfflinePlayer player,Double value){
         this.player = player;
         this.value = value;
     }
 
     @Override
     public int compareTo(PlayerValue o) {
-        return Integer.compare(this.value,o.value);
+        return Double.compare(this.value,o.value);
     }
 }
